realtime-sentiment-analysis/requirements.txt
azure-cognitiveservices-speech
requests
azure.core
azure.ai.textanalytics
openai
dotenv